We've implemented the delete mode and done some rudimentary testing and it seems to be working well. We've also finxed up all but one of the issues that were causing compiler warnings as well as removing unused functions from utils.hpp / .cpp.

+++ tools/chunkEditor/src/editMode.cpp
@@ -741,7 +741,8 @@ void getBgCharFromUser(const yx chunkSize, editingState & edState)
 		    {
 		      bool foundACS {false};
 
-		      for(int aCSIter {}; aCSIter < aCSChars.size(); ++aCSIter)
+		      for(int aCSIter {}; (size_t)aCSIter < aCSChars.size();
+			  ++aCSIter)
 			{
 			  if(acs_map[charAtPos & A_CHARTEXT] ==
 			     aCSChars[aCSIter])
